Move optimize_jet.h to optimize_autodiff.h
